fix a number of redundant casts in genoviz_sdk/demos
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@4181 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ genoviz_sdk/demo/src/genoviz/tutorial/NarcissusParser.java
@@ -275,7 +275,7 @@ public class NarcissusParser implements ContentParser {
 		int token = theTokens.nextToken();
 		if (StreamTokenizer.TT_NUMBER == token) {
 			begin = (int) theTokens.nval;
-			token = (int) theTokens.nextToken();
+			token = theTokens.nextToken();
 			if (StreamTokenizer.TT_NUMBER == token) {
 				end = (int) theTokens.nval;
 				return new Range(begin, end);
@@ -302,7 +302,7 @@ public class NarcissusParser implements ContentParser {
 		int token = theTokens.nextToken();
 		if (StreamTokenizer.TT_NUMBER == token) {
 			r[0] = (int) theTokens.nval;
-			token = (int) theTokens.nextToken();
+			token = theTokens.nextToken();
 			if (StreamTokenizer.TT_NUMBER == token) {
 				r[1] = (int) theTokens.nval;
 			}
